Quiz/requirements.txt
amqp==5.2.0
asgiref==3.7.2
attrs==23.2.0
billiard==4.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django==4.2.10
django-cors-headers==4.3.1
django-cron==0.6.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.5
mysqlclient==2.2.1
pillow==10.2.0
prompt-toolkit==3.0.43
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.32.1
rest-framework-simplejwt==0.0.2
rpds-py==0.17.1
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
